+++ nine/math/Vector2fAcceptor.java
@@ -2,5 +2,5 @@ package nine.math;
 
 public interface Vector2fAcceptor
 {
-    void call(float x, float y);
+    void call(double x, double y);
 }
