+++ nine/math/Vector2fFunction.java
@@ -4,5 +4,5 @@ public interface Vector2fFunction
 {
     public static final Vector2fFunction identity = (x, y, action) -> action.call(x, y);
 
-    void call(float x, float y, Vector2fAcceptor acceptor);
+    void call(double x, double y, Vector2fAcceptor acceptor);
 }
